+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add decorated formatter
 - Add decorator for level color changing
 - Add decorator for level emoji prefixes
+- Add codecov badge
 
 ## [0.1.0]
 - Add .travis.yml for CI support